requirements.txt
alembic==1.0.7
Babel==2.6.0
bleach==3.1.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dominate==2.3.5
elasticsearch==6.3.1
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
-e git://github.com/maxcountryman/flask-login.git@848088a9fc6e8c9c418e8820e072cde6ac81dc00#egg=Flask_Login
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
redis==3.2.0
requests==2.21.0
rq==0.13.0
six==1.12.0
SQLAlchemy==1.2.17
urllib3==1.24.1
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
